refactor(studio): split the metrics panel's two readouts
MetricsPanel was 356 lines. Two surfaces come out, plus the pieces three files
now share:

- `MetricsOverviewStats` — the single-shard readout, pure presentation of one
  snapshot.
- `MetricsAggregateView` — the cross-shard rollup: a different question, a
  different data shape, and only present after a fan-out.
- `stat-card.tsx` and `metrics-format.ts` — `StatCard` plus the duration and
  hit-rate formatters, so the three surfaces cannot drift on how a value reads.

Suppression deleted. 79 reports tests green.

Refs #230.
